Change the manifest format.
The redesign has several goals:

* Make noop changes as fast as possible, including in cases where the
  tests are vendored into third-party repositories, so that the manifest
  can be updated before every test run without imposing a substantial
  performance burden.

* Make it possible for tools such a the lint to only operate on files
  that have changed since the last run, thus providing a considerable
  performance improvement to those tools.

The general approach taken is to hash every file in the repository and
store that hash in the manifest. Then expensive operations such as
parsing the file are only performed when the hash has changed. The
desire to use the manifest as input to the lint means that every
lintable file, not just tests, has to be included.

On my machine this read+hash process takes about 10% of the total
execution time, so there is clearly some performance left on the
table. In particular:

* The gitignore implementation is not very performant (this also takes
  about 10% of the runtime).

* Constructing objects for each test can be slow, and could be done
  lazilly.

* Reading a large JSON file is intrinsically slow.

* Python itself doesn't fully optimise things like walking directories
  or regexps, which leaves some performance on the table in other areas.

This list is not intended to be exhaustive.